Commit 1.0.10.1
Mandatory update, any node/wallet/peer that is not running this wallet will not be able to A) Sync, B) connect, and C) Mine. This is to force the chain on one network. This is the most brudish of tactics that I can think of, and if this doesn't work I'll have to do some advanced stuff.